% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanner.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Keywords}
\format{
An object of class \code{list} of length 16.
}
\usage{
keywords
}
\description{
A list of reserved keywords and their corresponding token types.
}
\keyword{datasets}
